A mod allowing you to play the PokeFlute to summon Mew.
Play the PokeFlute in the basement of the Cinnabar Island mansion clandestine genetics lab to trigger Mew encounters. Mew no longer responds once captured.

This change includes a minor change to Pokeball logic that resulted in catching a Ditto if Mew had used Transform upon capture. The first playtest was a little disappointing. ;)