Adds a NPC to trade a Kadabra for an Alakazam to the hotel in Celadon.
This was my first mod based on the helpful tutorial at https://github.com/pret/pokered/wiki/Adding-an-NPC-that-will-trade-your-own-pokemon-back-for-evolution-purposes. In the original game you could only evolve Alakazam by trading, so it's semi-authentic, but the XP bonus feels like cheating. It went from a super mid Kadabra without Confusion to my tank. Because it's not a tradeback NPC, you lose your Kadabra's moves, so I ended up adding TM29 to the Celadon department store's inventory so I could teach it again.