Added the ability to evolve Machoke with lemonade.
The item logic for Soda Pop and Lemonade was redirected into a new function, ItemUseSoda, that first attempts to use soda as an evolution stone if outside of battle. A new evolution method using Lemonade was added to Machoke's evo table. I prefer this method of evolution for the Pokemon that can only be evolved by trade. It's just the right amount of dumb.